Finished Basic.hpp I think. Added namespaces for InSDK variables.

+++ Dumper/DependencyManager.h
@@ -17,7 +17,7 @@ private:
 	struct IndexDependencyInfo
 	{
 		/* Counter incremented every time this element is hit during iteration, **if** the counter is less than the CurrentIterationIndex */
-		mutable uint8 IterationHitCounter = 0x0;
+		mutable uint64 IterationHitCounter = 0x0;
 
 		/* Indices of Objects required by this Object */
 		std::unordered_set<int32> DependencyIndices;
@@ -27,8 +27,8 @@ private:
 	/* List of Objects and their Dependencies */
 	std::unordered_map<int32, IndexDependencyInfo> AllDependencies;
 
-	/* Count to track how often the Dependency-List was iterated. Allows for up to 255 iterations of this list. */
-	mutable uint8 CurrentIterationHitCount = 0x0;
+	/* Count to track how often the Dependency-List was iterated. Allows for up to 2^64 iterations of this list. */
+	mutable uint64 CurrentIterationHitCount = 0x0;
 
 public:
 	DependencyManager() = default;
